fix(convex): rename admission-snapshot so the production push is legal
`convex/companyMonitoring/admission-snapshot.ts` shipped in #6470 and took
the production Convex deploy down on main:

    InvalidConfig: companyMonitoring/admission-snapshot.js is not a valid
    path to a Convex module. Path component admission-snapshot.js can only
    contain alphanumeric characters, underscores, or periods.

`evaluate_push` validates before pushing anything, so this one filename
rejects the WHOLE deploy — every unrelated convex/ change merged after it
stays out of production, the same blast radius as #6232.

The convex npm package does not check the rule locally; the only "no"
comes from the server, post-merge, where nothing gates. So the rename is
paired with the pre-merge guard that reproduces the rule offline, in the
existing gate-required hygiene test rather than a new workflow.
