+++ convex/companyMonitoring/admissionSnapshot.ts
